restructured files to follow best practices
